requirements.txt
pandas==1.5.3
scikit-learn==1.2.2
mlflow==2.3.2
wandb==0.15.0
dvc==2.10.2
streamlit==1.22.0
pyyaml==6.0